Rename project from inlinexkcd to comicbot
Package and naming changes:
- Rename package: me.ex0ns.inlinexkcd → me.ex0ns.comicbot
- Rename directory structure using git mv (preserves history)
- Update all package declarations and imports (16 files)
- Update build.sbt project name: InlineXKCD → ComicBot
- Update assembly artifact name: InlineComicBot-assembly → ComicBot-assembly
- Update Dockerfile artifact path
- Update docker-compose service name: inline-comic-bot → comic-bot
- Update logger names in logback.xml
- Update CLI usage text (inlinecomicbot → comicbot)
- Update documentation (README.md, CLAUDE.md)

Database changes:
- Rename database: comics → comicbot
- Rename collection: comics → strips
- Updated documentation to reflect: comicbot.strips and comicbot.groups

Frontend changes:
- Update title: "XKCD Bot" → "Comic Bot"
- Update description to reflect multi-comic source support
- Fix database documentation

This generalizes the project name from "inlinexkcd" to "comicbot"
to better reflect its multi-comic source architecture.
