+++ src/main/resources/logback.xml
@@ -11,6 +11,6 @@
     </root>
 
     <logger name="org.asynchttpclient.netty" level="WARN" />
-    <logger name="me.ex0ns.inlinexkcd.bot" level="INFO" />
+    <logger name="me.ex0ns.comicbot.bot" level="INFO" />
     <logger name="com.bot4s" level="WARN" />
 </configuration>
