+++ src/main/scala/me/ex0ns/comicbot/models/Comic.scala
@@ -1,10 +1,10 @@
-package me.ex0ns.inlinexkcd.models
+package me.ex0ns.comicbot.models
 
 
 import com.bot4s.telegram.api.RequestHandler
 import com.bot4s.telegram.methods.{ParseMode, SendMessage, SendPhoto}
 import com.bot4s.telegram.models.InputFile
-import me.ex0ns.inlinexkcd.helpers.StringHelpers._
+import me.ex0ns.comicbot.helpers.StringHelpers._
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
